Fix overflow caused by scale animation
The overflow was difficult to find with the usual method of * selector
with a background color antiquewhite and border with red color and the
horizontal scrollbar was not seen on all the case and that was
confusing. The reason was because the scaling animation without
overflow hidden on it's container, as it no issue with overflow after
finish animation because of no overflow after it made very
difficult to find the issue.

So, let't make this description to go through in the future if such
situation arise and we don't remember the solution.